requirements.txt
streamlit>=1.30.0
python-docx>=0.8.11
dashscope>=1.14.0
pandas>=2.0.0
Pillow>=10.0.0
mammoth>=1.6.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
